+++ agents-cli/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @inkeep/agents-cli
 
+## 0.53.13
+
+### Patch Changes
+
+- Updated dependencies [e915ef8]
+  - @inkeep/agents-core@0.53.13
+  - @inkeep/agents-manage-ui@0.53.13
+  - @inkeep/agents-sdk@0.53.13
+
+## 0.53.12
+
+### Patch Changes
+
+- 9f5204d: Use human readable file names when generating files with inkeep pull
+- Updated dependencies [0038a02]
+  - @inkeep/agents-manage-ui@0.53.12
+  - @inkeep/agents-core@0.53.12
+  - @inkeep/agents-sdk@0.53.12
+
+## 0.53.11
+
+### Patch Changes
+
+- @inkeep/agents-manage-ui@0.53.11
+- @inkeep/agents-core@0.53.11
+- @inkeep/agents-sdk@0.53.11
+
+## 0.53.10
+
+### Patch Changes
+
+- b2f3879: Enable inkeep pull to handle nullable values in triggers.
+- Updated dependencies [eacb0dc]
+- Updated dependencies [33780a8]
+- Updated dependencies [ff9e307]
+- Updated dependencies [901058e]
+  - @inkeep/agents-core@0.53.10
+  - @inkeep/agents-manage-ui@0.53.10
+  - @inkeep/agents-sdk@0.53.10
+
+## 0.53.9
+
+### Patch Changes
+
+- Updated dependencies [9a2d783]
+- Updated dependencies [27cd96b]
+- Updated dependencies [9a2d783]
+- Updated dependencies [603d7a8]
+  - @inkeep/agents-core@0.53.9
+  - @inkeep/agents-manage-ui@0.53.9
+  - @inkeep/agents-sdk@0.53.9
+
+## 0.53.8
+
+### Patch Changes
+
+- Updated dependencies [7e60710]
+- Updated dependencies [a879934]
+  - @inkeep/agents-manage-ui@0.53.8
+  - @inkeep/agents-core@0.53.8
+  - @inkeep/agents-sdk@0.53.8
+
+## 0.53.7
+
+### Patch Changes
+
+- Updated dependencies [aa37d3f]
+- Updated dependencies [54985c0]
+  - @inkeep/agents-core@0.53.7
+  - @inkeep/agents-manage-ui@0.53.7
+  - @inkeep/agents-sdk@0.53.7
+
 ## 0.53.6
 
 ### Patch Changes